-move marginal_cost calculation to base DSM class -fix naive DSM strategy
